Add more information to schedule list, daily upcoming speedruns, allow connecting over ipv4
The list now includes a lot more information, like the runner and time estimate, as well as better formatting the platform as a badge. Checkboxes are bigger and easier to click, plus there's a new "show only interested" checkbox, and header-like banners whenever a speedrun is on a new day, like the official schedule has.

Now at midnight a message will be sent notifying of speedruns scheduled for that day that you're interested in.

The site now binds on 0.0.0.0 so it will be available over ipv4.