Make the upcoming speedruns clickable
Now clicking on a speedrun that's upcoming will scroll to that run in the table and highlight it for a second.

I also removed the runner information, don't need to show everything in those banners and the runner felt a little unnecessary for this, it's in the table which is now just a click away.