Update everything in relation to the current time as time changes
Everything that compared to a time now compares to a `now` store which is the current date, updated every second, so countdowns, the upcoming banner, and hidden past runs will be accurate no matter when you loaded the page.

This also makes it easier to test the ui as `now` can be set to any arbitrary date and the entire page pretends it's that time.

I also made the schedule refresh every 15 minutes, matching the refresh time on the backend, so new runs will show up.