Notify when runs were removed from the schedule
This just checks for any runs that are no longer on the new schedule and sends a notification with the same style as the other 'new runs' message. Interests periodically check for runs that are no longer on the schedule to remove them.

Also fixed a bug where loading the schedule wouldn't have valid date objects and if a notification needed to be sent immediately after loading but before more than one refresh the date formatting would throw errors and notifications wouldn't be sent.