Attempt to compensate for sudden schedule changes
The schedule can change and times can move, if a schedule change happened and a run that wasn't scheduled to start for another 10 minutes actually just started there should be a notification, so it will now check if a run has already started in addition to checking if it's starting soon, so we don't miss things accidentally.

Also made the refresh time check more aggressively to avoid being too late. It now checks every 5 minutes instead of ever 15.