Notify of recent runs only up to a point, blockquote upcoming runs message
If the user checked a speedrun from a previous day it'd send a message saying it's starting soon, because it's in the past. There is now a differentiation between runs that have recently started and ones that are soon to start. If a run is longer than 30 minutes in the past then it won't send any messages.