Dependency updates, fix the transition from no event to an event
Updated many UI dependencies. The rel change on the links in App.svelte are working around a bug with svelte (#7994 on their tracker).

The way the site tells if an event is on the schedule is by checking it was able to pull an event name from the page title (if the first `<h1>` ends with "schedule"), but the variables referenced were wrong (using the previously cached title) so it was never realizing a new event is scheduled.