Reset interests when a new event is found
The previous commit made it so the interests would reset when the server loads, but if it's kept online for a long time (between events) it would need to restart to clear out the old interests. This introduces a rudimentary event handler system on the schedule, that emits whenever a new event is found, interests will watch for that and reset its state.